+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.2 (pre-release)
 * [Enhancement] Some fixes in French localization.
+* [Fix #33] Allow to disable the vessel overlay.
 
 # 2.1 (March, 3rd, 2020)
 * [Change] Switch to C# compiler v6.0.